requirements.txt
Django
requests
nltk
pyspark
pandas
gunicorn
django-cors-headers
djangorestframework
